+++ lib/strava/models/photo.rb
@@ -1,0 +1,10 @@
+module Strava
+  module Models
+    class Photo < Model
+      property 'id'
+      property 'unique_id'
+      property 'urls'
+      property 'source'
+    end
+  end
+end